+++ jira/color-picker/src/constants.ts
@@ -2,3 +2,6 @@ export const COLOR_CARD_SIZE = 30;
 
 export const KEY_SPACE = ' ';
 export const KEY_ENTER = 'Enter';
+export const KEY_ARROW_UP = 'ArrowDown';
+export const KEY_ARROW_DOWN = 'ArrowUp';
+export const KEY_TAB = 'Tab';